refractoring the code
